+++ packages/arch/packages/input/src/Input.js
@@ -3,18 +3,18 @@
 import { jsx } from '@emotion/core';
 import { forwardRef } from 'react';
 
-import { buttonAndInputBase } from '@arch-ui/common';
+import { uniformHeight } from '@arch-ui/common';
 import { colors } from '@arch-ui/theme';
-import { alpha } from '@arch-ui/color-utils';
+// import { alpha } from '@arch-ui/color-utils';
 
 // Basic Input
 // ------------------------------
 
 export const inputStyles = (props: InputProps = {}) => ({
-  ...buttonAndInputBase,
+  ...uniformHeight,
   backgroundColor: props.disabled ? colors.N10 : 'white',
   borderColor: colors.N20,
-  boxShadow: 'inset 0 1px 1px rgba(0, 0, 0, 0.075)',
+  // boxShadow: 'inset 0 1px 1px rgba(0, 0, 0, 0.075)',
   color: 'inherit',
   width: '100%',
 
@@ -24,25 +24,26 @@ export const inputStyles = (props: InputProps = {}) => ({
   },
   ':focus': {
     borderColor: colors.primary,
-    boxShadow: `inset 0 1px 1px rgba(0, 0, 0, 0.075),
-        0 0 0 3px ${alpha(colors.primary, 0.2)}`,
+    // boxShadow: `inset 0 1px 1px rgba(0, 0, 0, 0.075), 0 0 0 3px ${alpha(colors.primary, 0.2)}`,
     outline: 0,
   },
   '&[disabled]': {
     borderColor: colors.N15,
-    boxShadow: 'none',
+    // boxShadow: 'none',
     backgroundColor: colors.N05,
   },
   ...(props.isMultiline
     ? {
         lineHeight: 'inherit',
-        height: 'auto',
+        height: 100,
+        resize: 'vertical',
       }
     : undefined),
 });
 
 type InputProps = { isMultiline?: boolean, disabled?: boolean };
 export const Input = forwardRef<InputProps, any>((props: InputProps, ref) => {
-  const Component = props.isMultiline ? 'textarea' : 'input';
-  return <Component ref={ref} css={inputStyles(props)} {...props} />;
+  const { isMultiline, ...inputProps } = props;
+  const Component = isMultiline ? 'textarea' : 'input';
+  return <Component ref={ref} css={inputStyles(props)} {...inputProps} />;
 });