+++ packages/arch/packages/modal-utils/src/index.js
@@ -4,6 +4,7 @@ export {
   SlideUp,
   SlideDown,
   SlideInHorizontal,
+  SpringDown,
   ZoomInDown,
   TransitionProvider,
   withTransitionState,