+++ packages/build-field-types/__fixtures__/common-dependency-two-entrypoints/src/sum.js
@@ -1,0 +1,3 @@
+import { identity } from './identity';
+
+export let sum = (a, b) => identity(a + b);