+++ packages/fields/src/types/Checkbox/views/Cell.js
@@ -1,0 +1,13 @@
+// @flow
+
+import React, { Component } from 'react';
+import type { CellProps } from '../../../types';
+
+type Props = CellProps<boolean>;
+
+export default class CheckboxCellView extends Component<Props> {
+  render() {
+    const { data } = this.props;
+    return <input type="checkbox" checked={data} disabled />;
+  }
+}