+++ packages/fields/types/Content/blocks/heading.js
@@ -1,8 +1,0 @@
-const path = require('path');
-const paragraph = require('./paragraph');
-
-module.exports = {
-  type: 'heading',
-  viewPath: path.join(__dirname, '../views/blocks/heading'),
-  dependencies: [paragraph],
-};