+++ packages/keystone/bin/cli.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+
 const arg = require('arg');
 const path = require('path');
 const globby = require('globby');